Add vmuser to sudoers with NOPASSWD (#15)
`vmuser` is already in the wheel group, but wheel still prompts for
password. I think this is nicer than setting
`security.sudo.wheelNeedsPassword` which changes more than just the
behavior of the `vmuser`.

We may also want to set up the same behavior under `security.sudo-rs`,
`security.doas`, and `security.please` in the future.