Don't use list for ssh command (#26)
I found this a bit confusing since it seems to suggest that the command
is not passed into a shell, but in both the test implementation and real
implementation we *do* pass this into a shell. I think it makes more
sense to treat the command as a single `Text` and do the concatenation
at the very outside where we receive it from *our* argv.