Use `Hostname` newtype
I was originally adding an extra check for `'` characters in
`updateVmHostsEntryImpl` since this is being concatenated into a shell.
However, this was just for extra sanity since we already check that the
hostname passed in was valid higher up in the call chain.

But, from this function's perspective, this isn't guaranteed since the
parameter is just a `Text`.

By adding a newtype wrapper around `Hostname` and not exporting the
constructor we can safely just concatenate the hostname in the ssh shell
without worrying that this function is somehow called in the future with
a shellism. Just like how we don't need to escape the concatenated ipv4.