Move up command into its own module (#60)
This PR just moves code around.

We also have an `UpSpec` module (not `Commands.UpSpec`). Which feels a
bit inconsistent, since the implementation of the `up` command is in
`Commands.Up`. I thought this was better though, since the `UpSpec`
module does not test the interface of `Commands.Up`, but rather the `up`
command through the command line interface. Happy to change that though.
:shrug: